extension/requirements.txt
vscode-ext